@@ -61,6 +61,7 @@
     "howler": "^2.2.3",
     "lodash": "^4.17.21",
     "rxjs": "~7.5.5",
+    "swiper": "^11.1.11",
     "tslib": "^2.2.0",
     "zone.js": "~0.14.10"
   },
@@ -80,9 +81,9 @@
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "eslint": "^7.6.0",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "eslint": "^7.26.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jsdoc": "30.7.6",
     "eslint-plugin-prefer-arrow": "1.2.2",
