@@ -1,7 +1,7 @@
 import { Puzzle } from './puzzle.model';
 import { UserPuzzle } from './user-puzzles.model';
 
-export type PlanTypes = 'warmup' | 'plan5' | 'plan10' | 'plan20' | 'plan30' | 'backToCalm';
+export type PlanTypes = 'warmup' | 'plan5' | 'plan10' | 'plan20' | 'plan30' | 'backToCalm' | 'custom';
 
 
 
