@@ -1,4 +1,4 @@
-import { Component, OnInit, OnChanges, SimpleChanges, ViewChild, ElementRef, AfterViewInit, Input } from '@angular/core';
+import { Component, OnInit, OnChanges, SimpleChanges, ViewChild, ElementRef, AfterViewInit, Input, OnDestroy } from '@angular/core';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 
@@ -23,8 +23,11 @@ import { PlanService } from '@services/plan.service';
   templateUrl: './activity-chart.component.html',
   styleUrls: ['./activity-chart.component.scss'],
 })
-export class ActivityChartComponent implements OnInit, AfterViewInit {
+export class ActivityChartComponent implements OnInit, AfterViewInit, OnChanges, OnDestroy {
 
+  // TODO: recibir planes por input, en el onInit se verifica si ya tiene planes para hacer el setData()
+  // y se activa una bantedera, en caso tal de que no tenga datos todavia, y en el input set
+  // se verifica esa bandera, si esta activa es porque ya se cargo el onInit y el grafico esta listo para ser poblado
 
   @ViewChild('matrixChart') matrixChartRef: ElementRef;
 
@@ -45,6 +48,7 @@ export class ActivityChartComponent implements OnInit, AfterViewInit {
       LineElement, Title, Tooltip, Filler,
       TimeScale);
 
+    // TODO: esto se debe recibir por el Input
     this.planService.getPlansHistoryState().pipe(takeUntil(this.destroy$)).subscribe(data => {
       this.setData(data);
     });
