@@ -1,6 +1,8 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, ValidatorFn, ValidationErrors } from '@angular/forms';
 
+import { ModalController } from '@ionic/angular';
+
 import { AppPuzzlesThemes, AppPuzzleThemesGroup } from '@models/app.models';
 import { Block } from '@models/plan.model';
 
@@ -22,13 +24,13 @@ export class BlockSettingsComponent implements OnInit {
   puzzlesGroupsThemes: AppPuzzleThemesGroup[] = [];
   form: FormGroup;
 
-  private dashObligatoryDuration: boolean;
 
 
   constructor(
     private formBuilder: FormBuilder,
     private appService: AppService,
-    public uiService: UiService
+    public uiService: UiService,
+    private modalController: ModalController
   ) {
     this.puzzlesGroupsThemes = this.appService.getThemesPuzzle;
 
@@ -42,29 +44,22 @@ export class BlockSettingsComponent implements OnInit {
     return this.form.get('puzzlesCount');
   }
 
-  get themesField() {
-    return this.form.get('themes');
+  get themeField() {
+    return this.form.get('theme');
   }
 
-  get eloStartField() {
-    return this.form.get('eloStart');
+  get eloLevelField() {
+    return this.form.get('eloLevel');
   }
 
-  get eloEndField() {
-    return this.form.get('eloEnd');
+  get puzzleTimeField() {
+    return this.form.get('puzzleTime');
   }
 
-
-  get obligatoryDuration(): boolean {
-    return this.dashObligatoryDuration;
+  get goshPuzzleTimeField() {
+    return this.form.get('goshPuzzleTime');
   }
 
-
-  @Input()
-  set obligatoryDuration(value: boolean) {
-    this.dashObligatoryDuration = value;
-    this.updateFormValidators();
-  }
 
 
 
@@ -75,57 +70,28 @@ export class BlockSettingsComponent implements OnInit {
 
   buildForm() {
     this.form = this.formBuilder.group({
-      time: [{ value: null, disabled: true }, Validators.min(3)],
-      durationTime: false,
-      puzzlesCount: [{ value: null, disabled: true }, Validators.min(1)],
-      durationCount: false,
-      // adicionar validacion minimo 800 y maximo 3000
-      eloStart: [800, Validators.compose([Validators.required, Validators.min(800), Validators.max(2900)])],
-      eloEnd: [3000, Validators.compose([Validators.required, Validators.min(900), Validators.max(3000)])],
-      themes: 'all',
+      time: [300, Validators.required],
+      puzzlesCount: [],
+      eloLevel: [0],
+      theme: 'all',
       openingFamily: '',
       puzzleTime: [60, Validators.required],
       nextPuzzleImmediately: true,
       showPuzzleSolution: true,
       goshPuzzle: false,
       goshPuzzleTime: [{ value: 30, disabled: true }]
-    }, {
-      validators: [this.rangeDifferenceValidator, this.durationValidator()]
     });
 
     this.form.get('goshPuzzle').valueChanges.subscribe(() => {
       this.toggleFieldBasedOnBoolean('goshPuzzle', 'goshPuzzleTime');
     });
-    this.form.get('durationTime').valueChanges.subscribe(() => {
-      this.toggleFieldBasedOnBoolean('durationTime', 'time');
-    });
-    this.form.get('durationCount').valueChanges.subscribe(() => {
-      this.toggleFieldBasedOnBoolean('durationCount', 'puzzlesCount');
-    });
   }
 
-  rangeDifferenceValidator(group: FormGroup): { [key: string]: boolean } | null {
-    const eloStart = group.get('eloStart').value;
-    const eloEnd = group.get('eloEnd').value;
-    if (eloEnd - eloStart < 100) {
-      return { rangeTooSmall: true };
-    }
-    return null;
+  formatPin(value: number) {
+    return value > 0 ? `+${value}` : `${value}`;
   }
 
-  durationValidator(): ValidatorFn {
-    return (group: FormGroup): ValidationErrors | null => {
-      if (this.obligatoryDuration) {
-        const timeValue = group.get('time').value;
-        const puzzlesCountValue = group.get('puzzlesCount').value;
 
-        if (!timeValue && !puzzlesCountValue) {
-          return { durationRequired: true };
-        }
-      }
-      return null;
-    };
-  }
 
   toggleFieldBasedOnBoolean(booleanControlName: string, targetControlName: string) {
     const booleanControl = this.form.get(booleanControlName);
@@ -146,12 +112,6 @@ export class BlockSettingsComponent implements OnInit {
     if (!this.form) {
       return; // Si el formulario aún no está inicializado, simplemente regresa
     }
-
-    if (this.obligatoryDuration) {
-      this.form.setValidators([this.rangeDifferenceValidator, this.durationValidator()]);
-    } else {
-      this.form.setValidators([this.rangeDifferenceValidator]);
-    }
     this.form.updateValueAndValidity();
   }
 
@@ -162,12 +122,21 @@ export class BlockSettingsComponent implements OnInit {
     if (this.form.invalid) {
       return;
     }
-
-    const themesValue = this.form.value.themes;
     // emit new block
+    const newBlock: Block = {
+      ...this.form.value,
+      color: this.color,
+      time: this.form.value.time === 0 ? -1 : this.form.value.time,
+      puzzleTimes: {
+        total: this.form.value.puzzleTime,
+        warningOn: this.form.value.puzzleTime / 2,
+        dangerOn: this.form.value.puzzleTime / 4
+      }
+    };
+    this.modalController.dismiss(newBlock);
+  }
 
-    this.newBlock.emit({ ...this.form.value, color: this.color, themes: themesValue !== 'all' ? [themesValue] : [] });
-
-
+  close() {
+    this.modalController.dismiss();
   }
 }
