@@ -46,7 +46,12 @@ import { createUid } from '@utils/create-uid';
 })
 export class BoardPuzzleComponent implements OnInit {
 
+  @Output() puzzleCompleted = new EventEmitter<Puzzle>();
+  @Output() puzzleFailed = new EventEmitter<Puzzle>();
+  @Output() puzzleEndByTime = new EventEmitter<Puzzle>();
+
   puzzle: Puzzle;
+  isPlaying = false;
 
   currentMoveNumber = 0;
   arrayFenSolution = [];
@@ -64,17 +69,10 @@ export class BoardPuzzleComponent implements OnInit {
   timerUnsubscribe$ = new Subject<void>();
 
   timeUsed = 0;
-
   goshPuzzleTime = 0;
-
-
   board;
   chessInstance = new Chess();
   piecePathKingTurn = '';
-
-  @Output() puzzleCompleted = new EventEmitter<Puzzle>();
-  @Output() puzzleFailed = new EventEmitter<Puzzle>();
-  @Output() puzzleEndByTime = new EventEmitter<Puzzle>();
 
   constructor(
     private renderer: Renderer2,
@@ -91,7 +89,6 @@ export class BoardPuzzleComponent implements OnInit {
 
   @Input() set setForceStopTimer(data: boolean) {
     if (data) {
-      // por que no se detiene el timer?
       this.stopTimer();
     }
   }
@@ -137,6 +134,8 @@ export class BoardPuzzleComponent implements OnInit {
 
     // ejecutar primera jugada
     this.puzzleMoveResponse();
+
+    this.isPlaying = true;
 
     // se valida si el puzzle tiene un tiempo limite para resolverlo
     if (this.puzzle?.times?.total) {
@@ -253,7 +252,6 @@ export class BoardPuzzleComponent implements OnInit {
     let endSquare;
     this.board.enableSquareSelect((event) => {
 
-
       const ctrKeyPressed = event.mouseEvent.ctrlKey;
       const shiftKeyPressed = event.mouseEvent.shiftKey;
       const altKeyPressed = event.mouseEvent.altKey;
@@ -284,7 +282,6 @@ export class BoardPuzzleComponent implements OnInit {
         } else if (ctrKeyPressed) {
           arrowType = { ...arrowType, class: 'arrow-red' };
         }
-
 
         this.board.addArrow(arrowType, startSquare, endSquare);
       }
@@ -327,7 +324,6 @@ export class BoardPuzzleComponent implements OnInit {
           this.board.addMarker(myOwnMarker, event.square);
         }
 
-
       }
     });
 
@@ -346,7 +342,6 @@ export class BoardPuzzleComponent implements OnInit {
         this.board.removeMarkers(marker.type, square);
       }
     });
-
 
   }
 
@@ -403,6 +398,7 @@ export class BoardPuzzleComponent implements OnInit {
             firstMoveSquaresHighlight: [this.arrayMovesSolution[0].slice(0, 2), this.arrayMovesSolution[0].slice(2, 4)]
           });
           this.stopTimer();
+          this.isPlaying = false;
         }
       } else {
         this.time++;
@@ -465,6 +461,7 @@ export class BoardPuzzleComponent implements OnInit {
         firstMoveSquaresHighlight: [this.arrayMovesSolution[0].slice(0, 2), this.arrayMovesSolution[0].slice(2, 4)]
       });
       this.stopTimer();
+      this.isPlaying = false;
     }
 
     // Actualiza el tablero después de un movimiento de enroque
@@ -493,6 +490,7 @@ export class BoardPuzzleComponent implements OnInit {
         firstMoveSquaresHighlight: [this.arrayMovesSolution[0].slice(0, 2), this.arrayMovesSolution[0].slice(2, 4)]
       });
       this.stopTimer();
+      this.isPlaying = false;
     } else {
 
       await new Promise<void>((resolve, reject) => {
