@@ -1,10 +1,15 @@
 import { TrainingComponent } from './training/training.component';
 import { TrainingMenuComponent } from './training-menu/training-menu.component';
-
+import { CustomTrainingComponent } from './custom-training/custom-training.component';
+import { PlanPlayedComponent } from './plan-played/plan-played.component';
+import { PlansHistorySlidesComponent } from './plans-history-slides/plans-history-slides.component';
 
 export const CONTAINERS: any[] = [
     TrainingComponent,
-    TrainingMenuComponent
+    TrainingMenuComponent,
+    CustomTrainingComponent,
+    PlanPlayedComponent,
+    PlansHistorySlidesComponent
 ];
 
 export const ENTRY_CONTAINERS: any[] = [
