@@ -3,10 +3,12 @@ import { AuthEffects } from './auth.effects';
 import { CoordinatesEffects } from './coordinates.effects';
 import { UserPuzzlesEffects } from './user-puzzles.effects';
 import { PlanEffects } from './plan.effects';
+import { PlaneElosEffects } from './plan-elos.effects';
 
 export const EFFECTS: any[] = [
     AuthEffects,
     CoordinatesEffects,
     UserPuzzlesEffects,
-    PlanEffects
+    PlanEffects,
+    PlaneElosEffects
 ];
