@@ -17,7 +17,7 @@ import { setPlan, requestSavePlan, requestGetPlans } from '@redux/actions/plans.
 
 // Selectors
 import { getPlan } from '@redux/selectors/plan.selectors';
-import { getAllPlansHistory } from '@redux/selectors/plans-history.selectors';
+import { getPlansHistoryOrderByDate } from '@redux/selectors/plans-history.selectors';
 
 // services
 import { FirestoreService } from '@services/firestore.service';
@@ -37,13 +37,20 @@ export class PlanService {
   /**
    * Actions
    */
-  requestSavePlanAction(plan: Plan) {
 
+  setPlanAction(plan: Plan) {
+    this.store.dispatch(setPlan({ plan }));
+  }
+
+  requestSavePlanAction(plan: Plan) {
     // clear puzzles in blocks
-    plan.blocks = plan.blocks.map((block: Block) => {
-      block.puzzles = [];
-      return block;
-    });
+    plan = {
+      ...plan,
+      blocks: plan.blocks.map((block: Block) => ({
+        ...block,
+        puzzles: []
+      }))
+    };
     this.store.dispatch(requestSavePlan({ plan }));
   }
 
@@ -61,7 +68,7 @@ export class PlanService {
   /** STATE OBSERVABLES */
 
   getPlansHistoryState(): Observable<Plan[]> {
-    return this.store.select(getAllPlansHistory);
+    return this.store.select(getPlansHistoryOrderByDate);
   }
 
   /**
@@ -70,20 +77,15 @@ export class PlanService {
    * @param time in seconds (-1 for infinite)
    * */
   newPlan(blocks: Block[], planType: PlanTypes, time = -1): Promise<Plan> {
-
     return new Promise((resolve, reject) => {
-
       const plan: Plan = {
         uid: createUid(),
         blocks,
         planType,
         createdAt: new Date().getTime(),
       };
-
-      this.store.dispatch(setPlan({ plan }));
-
+      this.setPlanAction(plan);
       resolve(plan);
-
     });
   }
 
@@ -91,22 +93,17 @@ export class PlanService {
    * Get the current plan
    * */
   getPlan(): Promise<Plan> {
-
     return new Promise((resolve, reject) => {
-
       this.store.select(getPlan).subscribe((plan: Plan) => {
         resolve(plan);
       });
-
     });
-
   }
-
 
   /**
    * Save the plan
    */
-  savePlan(plan: Plan): Promise<string> {
+  savePlan(plan: Plan) {
     return this.firestoreService.savePlan(plan);
   }
 }
