@@ -3,6 +3,7 @@ import { StringToFlagPipe } from './string-to-flag.pipe';
 import { GameClockPipe } from './game-clock.pipe';
 import { FlagUrlPipe } from './flag-url.pipe';
 import { SecondsToMinutesSecondsPipe } from './seconds-to-minutes-seconds.pipe';
+import { AgoPipe } from './ago.pipe';
 
 
 export const PIPES = [
@@ -10,5 +11,6 @@ export const PIPES = [
     StringToFlagPipe,
     GameClockPipe,
     FlagUrlPipe,
-    SecondsToMinutesSecondsPipe
+    SecondsToMinutesSecondsPipe,
+    AgoPipe
 ];
